# feat: use number set to speed up number parsing