```
name                           old time/op    new time/op    delta
Decoder_Float64/floats.json-4    23.4µs ± 2%    21.4µs ± 2%  -8.47%  (p=0.000 n=15+14)

name                           old alloc/op   new alloc/op   delta
Decoder_Float64/floats.json-4     0.00B          0.00B         ~     (all equal)

name                           old allocs/op  new allocs/op  delta
Decoder_Float64/floats.json-4      0.00           0.00         ~     (all equal)
```